requirement.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.2
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-filter==2.2.0
django-guardian==2.1.0
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
httpie==1.0.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
Pillow==7.0.0
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
